Reorganize viewmodel placement
