+++ 3DPoC/POC3D/POC3D.ViewModel.Tests/NodeViewModelTests.cs
@@ -1,5 +1,6 @@
 ﻿using NUnit.Framework;
 using POC3D.Model;
+using POC3D.ViewModel.Implementation;
 using POC3D.ViewModel.Tests.Helper;
 using System;
 using System.Collections.Generic;
