+++ 3DPoC/POC3D/POC3D.ViewModel/Calculations/MatrixHelper.cs
@@ -1,7 +1,7 @@
 ﻿using System.Linq;
 using System.Windows.Media.Media3D;
 using POC3D.Model;
-using POC3D.ViewModel;
+using POC3D.ViewModel.Implementation;
 
 namespace POC3D.ViewModel.Calculations
 {
