+++ 3DPoC/POC3D/POC3D.ViewModel/Calculations/ProblemCalculationViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using POC3D.ViewModel.Base;
+using POC3D.ViewModel.Implementation;
 using System.Threading.Tasks;
 using System.Windows;
