+++ 3DPoC/POC3D/POC3D.ViewModel/Commands/AddForceCommand.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using POC3D.ViewModel.Implementation;
+using POC3D.ViewModel.InterfaceManagement;
+using System;
 using System.ComponentModel;
 using System.Linq;
 using System.Windows.Input;
