+++ 3DPoC/POC3D/POC3D.ViewModel/Commands/AddNodeCommand.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using POC3D.ViewModel.Implementation;
+using System;
 using System.Windows.Input;
 
 namespace POC3D.ViewModel.Commands
