+++ 3DPoC/POC3D/POC3D.ViewModel/Geometry/ResultNodeGeometryViewmodel.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using POC3D.ViewModel.Implementation;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
