+++ 3DPoC/POC3D/POC3D.ViewModel/InterfaceManagement/CameraViewModel.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Windows.Media.Media3D;
 
-namespace POC3D.ViewModel
+namespace POC3D.ViewModel.InterfaceManagement
 {
     public class CameraViewModel : Observable
     {
