+++ 3DPoC/POC3D/POC3D.ViewModel/InterfaceManagement/ForceManagementViewModel.cs
@@ -3,8 +3,9 @@ using System.Linq;
 using System.Windows.Input;
 using POC3D.ViewModel.Base;
 using POC3D.ViewModel.Commands;
+using POC3D.ViewModel.Implementation;
 
-namespace POC3D.ViewModel
+namespace POC3D.ViewModel.InterfaceManagement
 {
     public class ForceManagementViewModel : Observable
     {
