+++ 3DPoC/POC3D/POC3D.ViewModel/InterfaceManagement/InterfaceControlViewModel.cs
@@ -4,8 +4,9 @@ using System.Windows;
 using System.Windows.Input;
 using POC3D.ViewModel.Base;
 using POC3D.ViewModel.Commands;
+using POC3D.ViewModel.Implementation;
 
-namespace POC3D.ViewModel
+namespace POC3D.ViewModel.InterfaceManagement
 {
     public class InterfaceControlViewModel : Observable
     {
@@ -29,7 +30,7 @@ namespace POC3D.ViewModel
         {
             _problemViewModel = problemViewModel;
             _cameraViewModel = cameraViewModel;
-            
+
             ElementManagementViewModel = new ElementManagementViewModel(_problemViewModel);
             ForceManagementViewModel = new ForceManagementViewModel(_problemViewModel);
             NodeManagementViewModel = new NodeManagementViewModel(_problemViewModel);
