+++ 3DPoC/POC3D/POC3D.ViewModel/InterfaceManagement/NodeManagementViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using POC3D.ViewModel.Commands;
+using POC3D.ViewModel.Implementation;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -6,7 +7,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Input;
 
-namespace POC3D.ViewModel
+namespace POC3D.ViewModel.InterfaceManagement
 {
     public class NodeManagementViewModel
     {
