+++ 3DPoC/POC3D/POC3D.ViewModel/MainViewModel.cs
@@ -1,4 +1,6 @@
 ﻿using POC3D.ViewModel.Base;
+using POC3D.ViewModel.Implementation;
+using POC3D.ViewModel.InterfaceManagement;
 
 namespace POC3D.ViewModel
 {
