+++ 3DPoC/POC3D/POC3D.ViewModel/POC3D.ViewModel.csproj
@@ -55,7 +55,8 @@
     <Compile Include="Base\SelectableViewModel.cs" />
     <Compile Include="Calculations\ElementCalculationViewModel.cs" />
     <Compile Include="Calculations\ProblemCalculationViewModel.cs" />
-    <Compile Include="CameraViewModel.cs" />
+    <Compile Include="Implementation\ElementViewModel.cs" />
+    <Compile Include="InterfaceManagement\CameraViewModel.cs" />
     <Compile Include="Geometry\ResultElementGeometryViewModel.cs" />
     <Compile Include="Geometry\ElementGeometryViewModel.cs" />
     <Compile Include="Geometry\ForceGeometryViewModel.cs" />
@@ -70,22 +71,21 @@
     <Compile Include="Commands\DeleteElementCommand.cs" />
     <Compile Include="Commands\DeleteForceCommand.cs" />
     <Compile Include="Commands\DeleteNodeCommand.cs" />
-    <Compile Include="ElementViewModel.cs" />
-    <Compile Include="ForceViewModel.cs" />
+    <Compile Include="Implementation\ForceViewModel.cs" />
     <Compile Include="Calculations\ConstantMatrix.cs" />
     <Compile Include="Calculations\CorrespondenceMatrix.cs" />
     <Compile Include="Geometry\GraphicsHelper.cs" />
     <Compile Include="Geometry\NodeGeometryViewModel.cs" />
     <Compile Include="Geometry\ResultForceGeometryViewModel.cs" />
     <Compile Include="Geometry\ResultNodeGeometryViewModel.cs" />
-    <Compile Include="InterfaceControlViewModel.cs" />
+    <Compile Include="InterfaceManagement\InterfaceControlViewModel.cs" />
     <Compile Include="MainViewModel.cs" />
-    <Compile Include="MaterialViewModel.cs" />
-    <Compile Include="ElementManagementViewModel.cs" />
-    <Compile Include="ForceManagementViewModel.cs" />
-    <Compile Include="NodeManagementViewModel.cs" />
-    <Compile Include="NodeViewModel.cs" />
-    <Compile Include="ProblemViewModel.cs" />
+    <Compile Include="Implementation\MaterialViewModel.cs" />
+    <Compile Include="InterfaceManagement\ElementManagementViewModel.cs" />
+    <Compile Include="InterfaceManagement\ForceManagementViewModel.cs" />
+    <Compile Include="InterfaceManagement\NodeManagementViewModel.cs" />
+    <Compile Include="Implementation\NodeViewModel.cs" />
+    <Compile Include="Implementation\ProblemViewModel.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
